requirements.txt
streamlit
supabase
google-generativeai
python-dotenv
bcrypt